Convert aborted test and class in jar detection tests to new JUnit multiversion tests
